[Refactor] : type폴더, 파일 생성
