+++ client/src/pages/member/Mypage.tsx
@@ -11,8 +11,10 @@ import { Api } from '../../apis/customAPI';
 import IntroBox from '../../Components/member/IntroBox';
 import MyReview from '../../Components/member/MyReview';
 import useAxios from '../../hooks/useAxios';
-import { Iuser, UPDATE } from '../../reducers/userInfoReducer';
+import { UPDATE } from '../../reducers/userInfoReducer';
 import { RootState } from '../../store/Store';
+import { Ipost } from '../../type/Ipost';
+import { Iuser } from '../../type/Iuser';
 
 const Main = styled.div`
 	display: flex;
@@ -148,20 +150,6 @@ const UserWriting = styled.div`
 		}
 	}
 `;
-
-interface Ipost {
-	id: number;
-	subject: string;
-	title: string;
-	content?: string;
-	nickname: string;
-	email: string;
-	tag?: null;
-	voteCount?: number;
-	viewCount?: number;
-	createdAt?: string;
-	modifiedAt?: string;
-}
 
 function Mypage() {
 	const dispatch = useDispatch();
