+++ client/src/type/Ilogin.ts
@@ -1,0 +1,4 @@
+export interface Ilogin {
+	isLogin: boolean;
+	accessToken?: string;
+}
