+++ client/src/type/Ipost.ts
@@ -1,0 +1,14 @@
+export interface Ipost {
+	id: number;
+	subject: string;
+	title: string;
+	content?: string;
+	nickname: string;
+	email: string;
+	tag?: null;
+	voteCount?: number;
+	viewCount?: number;
+	createdAt?: string;
+	modifiedAt?: string;
+}
+export type Iposts = Ipost[];
