+++ client/src/type/Iuser.ts
@@ -1,0 +1,8 @@
+export interface Iuser {
+	id?: number;
+	nickname: string;
+	email?: string;
+	mbti: string;
+	img?: string;
+	badge?: null | string;
+}
